% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns_helper.R
\name{check_contents}
\alias{check_contents}
\title{Check a vector for required contents}
\usage{
check_contents(req, check)
}
\arguments{
\item{req}{vector of the required contents}

\item{check}{vector that must contain the required contents}
}
\value{
TRUE if the all required elements are within the check vector
otherwise an error is thrown.
}
\description{
This is an internal package function that is helpful to make sure
the conditions are met.
}
