% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns_data.R
\name{convert_nc_SEM}
\alias{convert_nc_SEM}
\title{Convert a NetCDF file into a data frame that SEM can use}
\usage{
convert_nc_SEM(nc_path)
}
\arguments{
\item{nc_path}{\describe{character path to the .nc file of met data, this
file must contain PAR, temp, VPD, and precip data}}
}
\value{
a data frame consisting of columns of the following:
\describe{
\item{time}{the day of the year}
\item{PAR}{incoming photosynthetically active radiation, umol/m2/s}
\item{temp}{air temperature, degrees C}
\item{VPD}{vapor pressure deficit, kPa}
\item{precip}{precipitation, mm}
}
}
\description{
Convert a NetCDF file into a data frame that SEM can use
}
