% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{createModel}
\alias{createModel}
\title{Create Model Fomulas}
\usage{
createModel(adjacency_matrix, variable_names, type = "list")
}
\arguments{
\item{adjacency_matrix}{the adjacency matrix.}

\item{variable_names}{a vector of variable names.}

\item{type}{determines the datatype of output, must be 'list' or 'string'}
}
\value{
a list/string that stores the relation fomula of each node.
}
\description{
Return an object which stores the dependence fomula, e.g. 'y ~ x1 + x2' of
each node with two different data types.
}
\examples{
adj_mat <- matrix(c(0,0,0,1,0,0,0,1,0), nrow=3, ncol=3, byrow=TRUE)
variable_names <- c('x1','x2','x3')

# get a list object
createModel(adj_matrix, variable_names)

# get a string object
createModel(adj_matrix, variable_names, type='string')
}
