% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinelist.R
\name{dvinelist}
\alias{dvinelist}
\title{Get List of Dvine Objects}
\usage{
dvinelist(
  fomula,
  adjacency_matrix,
  variable_names,
  data,
  uscale = FALSE,
  family_set = "parametric"
)
}
\arguments{
\item{fomula}{a list of model fomulas.}

\item{adjacency_matrix}{the adjacency matrix.}

\item{variable_names}{vector of variable names.}

\item{data}{x-scaled data, used to train the dvine model.}

\item{uscale}{default value: FALSE, used in \code{vinereg::vinereg}.}

\item{family_set}{used in \code{vinereg::vinereg}.}
}
\value{
a list of dvine objects and a list of quantile graphs
}
\description{
Get List of Dvine Objects
}
\examples{
#

}
