% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FME.R
\name{format_sensout}
\alias{format_sensout}
\title{Format the output returned by memc_sensrange or memc_sensfunc into long data format for easy plotting}
\usage{
format_sensout(obj)
}
\arguments{
\item{obj}{object returned by memc_sensrange}
}
\value{
the a long dataframe of the summary memc_sensrange
}
\description{
Format the output returned by memc_sensrange or memc_sensfunc into long data format for easy plotting
}
\seealso{
Other sensitivity: 
\code{\link{SEM_sensfunc}()},
\code{\link{SEM_sensrange}()}
}
\concept{sensitivity}
