% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adjM_name.R
\name{get_igraph_obj}
\alias{get_igraph_obj}
\title{Get Adjacency Matrix and Vairable Names}
\usage{
get_igraph_obj(edge_list, directed = TRUE)
}
\arguments{
\item{edge_list}{a list of 2 columns, directed edges pointing from column 1 to 2.}

\item{directed}{to tell if the graph is directed or not, default value is TRUE.}
}
\description{
Get an igraph object from a given edge list.
}
\examples{

df2_tr <- data('example_edges') # edge list

dag_tr <- get_igraph_obj(df2_tr)
}
