% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_graph.R
\name{get_subedgelist}
\alias{get_subedgelist}
\title{Separate Unconnected Graph}
\usage{
get_subedgelist(node_name, edgelist, directed = TRUE)
}
\arguments{
\item{node_name}{the name of the variable.}

\item{edgelist}{a list describing the edges from a graph.}

\item{directed}{true or false.}
}
\value{
an edge-list from a subgraph containing the given node i(if possible).
}
\description{
Want to know whether a graph is connected or not. If unconnected, I want to be able
to get a subgraph which contains a certain node we give.
}
\examples{
#

}
