% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adjM_name.R
\name{get_variable_names}
\alias{get_variable_names}
\title{Get Adjacency Matrix and Vairable Names}
\usage{
get_variable_names(dag_tr)
}
\arguments{
\item{dag_tr}{an igraph object.}
}
\description{
Get variable names from an igraph object.
}
\examples{
library(igraph)
data('example_edges')
df2_tr <- example_edges # edge list
edges <- as.matrix(df2_tr)
dag_tr <- graph_from_edgelist(edges, directed = TRUE)

get_variable_names(dag_tr)
}
