% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinelist.R, R/get_model_data.R
\name{getModelData}
\alias{getModelData}
\title{Get List of Dvine Objects and Quantile plots}
\usage{
getModelData(i, adjacency_matrix, variable_names, data)

getModelData(i, adjacency_matrix, variable_names, data)
}
\arguments{
\item{i}{an integral number taking from 1 to length(vairable_names).}

\item{adjacency_matrix}{the adjacency matrix.}

\item{variable_names}{vector of variable names.}

\item{data}{the data containing variable_names, can also be the u-scaled data.}

\item{fomula}{a list of model fomulas.}

\item{uscale}{default value: FALSE, used in \code{vinereg::vinereg}.}

\item{family_set}{used in \code{vinereg::vinereg}.}

\item{u_data}{optional parameter, used when uscale=TRUE}

\item{adj_matrix}{the adjacency matrix.}
}
\value{
a list of dvine objects and a list of quantile graphs

data of variables in the formula of i-th node.
}
\description{
Return sub-dataset of the original data that contains only the columns
used in the dependence formula.
}
\examples{
#

i <- 1
adj_mat <- matrix(c(0,0,0,1,0,0,0,1,0), nrow=3, ncol=3, byrow=TRUE)
variable_names <- c('x1','x2','x3')
data <- data.frame(x1=rnorm(10,0,1),x2=rnorm(10,0,1),x3=rnorm(10,0,1),x4=rnorm(10,0,1))

subdata <- getModelData(i,adj_matrix, variable_names,data)
}
