% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R
\docType{data}
\name{params_df}
\alias{params_df}
\title{A data.frame containing the SEM parameter values defined in Dietze and Matthes 2014.}
\format{
A data.frame of 2 columns and 33 rows. TODO might delete
\describe{
\item{parameter}{string of the SEM paramter name}
\item{value}{numeric value}
}
}
\usage{
params_df
}
\description{
A data.frame containing the SEM parameter values defined in Dietze and Matthes 2014.
}
\keyword{datasets}
