% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadoc.R
\docType{data}
\name{pools}
\alias{pools}
\title{A named numeric vector of the initial pool sizes for SEM.TODO might delete}
\format{
A data.frame of 2 columns and 33 rows.
}
\usage{
pools
}
\description{
A named numeric vector of the initial pool sizes for SEM.TODO might delete
}
\keyword{datasets}
