% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_table.R
\name{print_stat}
\alias{print_stat}
\title{Print Dvine based SEM in a table}
\usage{
print_stat(vc)
}
\arguments{
\item{vc}{a list of vinereg object.}
}
\value{
a table of all statistic scores.
}
\description{
Print Dvine based SEM in a table
}
