% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns_helper.R
\name{read_SEM_met}
\alias{read_SEM_met}
\title{Load and format the SEM meteorological input file}
\usage{
read_SEM_met(f)
}
\arguments{
\item{f}{character pathway of the meteorological file to read into R file should have the following contents
\describe{
\item{time}{date & time column YYYY-MM-DD HH:MM:SS}
\item{temp}{Air temperature, degrees C}
\item{precip}{Precipitation, mm}
\item{VPD}{Vapor pressure deficit, kPa}
\item{PAR}{Incoming photosynthetically active radiation, umol/m2/s}
}}
}
\value{
SEM paramter data frame with the new parameter values
}
\description{
This function helps handle the time column which can be affected by being
saved as a csv file.
}
\examples{
\donttest{
# Using the regular read.csv function can cause the problems with the time column. 
metdata <- read.csv(system.file("metdata/example_inputs.csv", package = "SEM"))
head(metdata$time)
class(metdata$time)

# Where as using the read_SEM_met function returns a the time column with the 
# correct date and time information. 
metdata <- read_SEM_met(system.file("metdata/example_inputs.csv", package = "SEM"))
head(metdata$time)
class(metdata$time)
}
}
