% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_model.R
\name{run_SEM}
\alias{run_SEM}
\title{Run SEM}
\usage{
run_SEM(
  inputs,
  X,
  param_df,
  DBH = 10,
  pest = c(phloem = 0, xylem = 0, leaf = 0, root = 0, stem = 0),
  pest.time = NULL,
  quiet = TRUE
)
}
\arguments{
\item{inputs}{named numeric vector containing the meteorological variables at a single time point
\describe{
\item{temp}{Air temperature, degrees C}
\item{precip}{Precipitation, mm}
\item{VPD}{Vapor pressure deficit, kPa}
\item{PAR}{Incoming photosynthetically active radiation, umol/m2/s}
\item{time}{as.POSIXct date fromat, must be every 30 min, TODO there needs to be some better way to handle the dates}
}
#' @param X named numeric vector containing the following
\describe{
\item{leaf}{kg/plant}
\item{wood}{kg/plant}
\item{root}{kg/plant}
\item{storage}{kg/plant}
\item{som}{soil organic matter, Mg/ha}
\item{soil_water}{m}
\item{stem_density}{stems/ha}
}}

\item{param_df}{dataframe containing the following
\describe{
\item{parameter}{character of the parameter name}
\item{value}{double parameter values}
}}

\item{DBH}{diameter at breast height default value set to 10}

\item{pest}{named vector for the pest impacts, default set to 0 pests
\describe{
\item{phloem}{phloem feaders}
\item{xylem}{xylem disrupters (bark beetle, canker, wilt, girdling)}
\item{leaf}{defoliators}
\item{root}{root rot}
\item{stem}{stem rot}
}}

\item{pest.time}{NULL or vector of the times to apply the pest disturbance to}

\item{quiet}{boolean default set to TRUE if set to FALSE will print date}
}
\value{
data frame of model results
}
\description{
Run SEM
}
