% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM-package.R
\docType{package}
\name{SEM-package}
\alias{SEM}
\alias{SEM-package}
\title{SEM: Simple Ecosystem Model}
\description{
This is an R package implementation of the SEM created by Dietze and Matthes. See Dietze, Michael C., and Jaclyn Hatala Matthes. 2014. A General Ecophysiological Framework for Modelling the Impact of Pests and Pathogens on Forest Ecosystems. Ecology Letters 17 (11): 1418–26 for the documenation manuscript and https://github.com/mdietze/PestED for the original model implementation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kdorheim.github.io/SEM/}
}

}
\author{
\strong{Maintainer}: Kalyn Dorheim \email{kalyn.dorheim@pnnl.com} (\href{https://orcid.org/0000-0001-8093-8397}{ORCID})

}
\keyword{internal}
