% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FME.R
\name{SEM_sensfunc}
\alias{SEM_sensfunc}
\title{Local sensitivity for SEM}
\usage{
SEM_sensfunc(
  pars,
  param_df,
  inputs,
  X,
  DBH,
  pest = c(phloem = 0, xylem = 0, leaf = 0, root = 0, stem = 0),
  pest.time = NULL,
  ...
)
}
\arguments{
\item{pars}{named numeric vector of the parameter values to be tested as part of the sensitivity analysis}

\item{param_df}{data frame of the parameters that are being held constant
\describe{
\item{parameter}{character of the parameter name}
\item{value}{double parameter values}
}}

\item{inputs}{named numeric vector containing the meteorological variables at a single time point
\describe{
\item{temp}{Air temperature, degrees C}
\item{precip}{Precipitation, mm}
\item{VPD}{Vapor pressure deficit, kPa}
\item{PAR}{Incoming photosynthetically active radiation, umol/m2/s}
\item{time}{as.POSIXct date fromat, must be every 30 min, TODO there needs to be some better way to handle the dates}
}}

\item{X}{named numeric vector containing the following
\describe{
\item{leaf}{kg/plant}
\item{wood}{kg/plant}
\item{root}{kg/plant}
\item{storage}{kg/plant}
\item{som}{soil organic matter, Mg/ha}
\item{soil_water}{m}
\item{stem_density}{stems/ha}
}}

\item{DBH}{diameter at breast height default value set to 10}

\item{pest}{default set to no disturbance otherwise is a named vector for the pest impacts}

\item{pest.time}{NULL or vector of the times to apply the pest disturbance to}

\item{...}{additional arguments passed to FME::sensRange such as num or dist}
}
\value{
results of the FME::sensRange
}
\description{
Estimate the local effect of a parameter on a SEM run model output
}
\seealso{
Other sensitivity: 
\code{\link{SEM_sensrange}()},
\code{\link{format_sensout}()}
}
\concept{sensitivity}
