% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns_helper.R
\name{update_params}
\alias{update_params}
\title{Update parameter values defined in the \code{params_df}}
\usage{
update_params(df, new)
}
\arguments{
\item{df}{data frame of the SEM parameter values}

\item{new}{named numeric vector of the parameter values to be updated}
}
\value{
SEM paramter data frame with the new parameter values
}
\description{
Update parameter values defined in the \code{params_df}
}
