% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvinelist.R
\name{dvinelist}
\alias{dvinelist}
\title{Get List of Dvine Objects}
\usage{
dvinelist(
  fomula,
  adjacency_matrix,
  variable_names,
  data,
  uscale = FALSE,
  family_set = "parametric"
)
}
\arguments{
\item{fomula}{a list of model fomulas.}

\item{adjacency_matrix}{the adjacency matrix.}

\item{variable_names}{vector of variable names.}

\item{data}{x-scaled data, used to train the dvine model.}

\item{uscale}{default value: FALSE, used in \code{vinereg::vinereg}.}

\item{family_set}{used in \code{vinereg::vinereg}.}
}
\value{
a list of dvine objects and a list of quantile graphs
}
\description{
Get List of Dvine Objects
}
\examples{
data('example_data')
assem <- example_data
assem<- as.data.frame(assem)
t1 <- as.data.frame(apply(assem, 2, scale))

data('example_edges')
df2_tr <- example_edges # edge list
edges <- as.matrix(df2_tr)
dag_tr <- graph_from_edgelist(edges, directed = TRUE)

adjacency_matrix <- get.adjacency(dag_tr, sparse = FALSE)
variable_names <- V(dag_tr)$name
fomula <- createModel(adjacency_matrix, variable_names,'list')
vc_list <- dvinelist(fomula, adjacency_matrix, variable_names,t1)

}
