% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adjM_name.R
\name{get_adj_matrix}
\alias{get_adj_matrix}
\title{Get Adjacency Matrix and Vairable Names}
\usage{
get_adj_matrix(edge_list, directed = TRUE)
}
\arguments{
\item{edge_list}{a list of 2 columns, directed edges pointing from column 1 to 2.}

\item{directed}{to tell if the graph is directed or not, default value is TRUE.}

\item{dag_tr}{an igraph object.}
}
\description{
Get the adjacency matrix. Within this function, man can obtain the adjacency
matrix from a listdescribing edges from a Graph.
}
\examples{
data('example_edges')
df2_tr <- example_edges # edge list

get_adj_matrix(df2_tr)
}
