% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_data.R
\name{getModelData}
\alias{getModelData}
\title{Obtain Data Used in the Formula}
\usage{
getModelData(i, adjacency_matrix, variable_names, data)
}
\arguments{
\item{i}{an integral number taking from 1 to length(vairable_names).}

\item{variable_names}{vector of variable names.}

\item{data}{the data containing variable_names, can also be the u-scaled data.}

\item{adj_matrix}{the adjacency matrix.}
}
\value{
data of variables in the formula of i-th node.
}
\description{
Return sub-dataset of the original data that contains only the columns
used in the dependence formula.
}
\examples{
library(igraph)
i <- 1

data('example_data')
assem <- example_data
assem<- as.data.frame(assem)
t1 <- as.data.frame(apply(assem, 2, scale))

data('example_edges')
df2_tr <- example_edges # edge list
edges <- as.matrix(df2_tr)
dag_tr <- graph_from_edgelist(edges, directed = TRUE)

adjacency_matrix <- get.adjacency(dag_tr, sparse = FALSE)
variable_names <- V(dag_tr)$name

x <- getModelData(i,adjacency_matrix, variable_names,t1)
}
