% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_data.R
\name{getModelData}
\alias{getModelData}
\title{Obtain Data Used in the Formula}
\usage{
getModelData(i, adjacency_matrix, variable_names, data)
}
\arguments{
\item{i}{an integral number taking from 1 to length(vairable_names).}

\item{variable_names}{vector of variable names.}

\item{data}{the data containing variable_names, can also be the u-scaled data.}

\item{adj_matrix}{the adjacency matrix.}
}
\value{
data of variables in the formula of i-th node.
}
\description{
Return sub-dataset of the original data that contains only the columns
used in the dependence formula.
}
\examples{
i <- 1
adj_mat <- matrix(c(0,0,0,1,0,0,0,1,0), nrow=3, ncol=3, byrow=TRUE)
variable_names <- c('x1','x2','x3')
data <- data.frame(x1=rnorm(10,0,1),x2=rnorm(10,0,1),x3=rnorm(10,0,1),x4=rnorm(10,0,1))

subdata <- getModelData(i,adj_matrix, variable_names,data)
}
