% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_table.R
\name{print_mod}
\alias{print_mod}
\title{Print Dvine based SEM in a table}
\usage{
print_mod(vc)
}
\arguments{
\item{vc}{a list of vinereg object.}
}
\value{
a table describing vine type between variables/ a table of all statistic scores.
}
\description{
Print Dvine based SEM in a table
}
\examples{
#

}
