% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_table.R
\name{print_mod}
\alias{print_mod}
\title{Print Dvine based SEM in a table}
\usage{
print_mod(vc)
}
\arguments{
\item{vc}{a list of vinereg object.}
}
\value{
a table describing vine type between variables.
}
\description{
Print Dvine based SEM in a table
}
\examples{
library(igraph)

data('example_data')
assem <- example_data
assem<- as.data.frame(assem)
t1 <- as.data.frame(apply(assem, 2, scale))

data('example_edges')
df2_tr <- example_edges # edge list
edges <- as.matrix(df2_tr)
dag_tr <- graph_from_edgelist(edges, directed = TRUE)

adjacency_matrix <- get.adjacency(dag_tr, sparse = FALSE)
variable_names <- V(dag_tr)$name
fomula <- createModel(adjacency_matrix, variable_names,'list')
vc <- dvinelist(fomula, adjacency_matrix, variable_names,t1)
print_mod(vc)
}
