% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_sem.R
\name{qplot_sem}
\alias{qplot_sem}
\title{Get Quantile Plots from a Gaussia SEM}
\usage{
qplot_sem(fitted_model, n)
}
\arguments{
\item{fitted_model}{The fitted gaussian structural model, a lavaan object.}

\item{n}{the number of points between 0 and 1 that will be generated,
with the default value 100.}
}
\description{
Get Quantile Plots from a Gaussia SEM
}
\examples{
# example code


}
