% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_sem.R
\name{qplot_sem}
\alias{qplot_sem}
\title{Get Quantile Plots from a Gaussia SEM}
\usage{
qplot_sem(
  fitted_sem,
  dat,
  alpha = c(0.1, 0.5, 0.9),
  adjacency_matrix,
  variable_names,
  fomula
)
}
\arguments{
\item{fitted_sem}{The fitted gaussian structural model, a lavaan object.}

\item{dat}{original dataset used to generate sem.}

\item{alpha}{default value is (0.1, 0.5, 0.9).}

\item{adjacency_matrix}{the adjacency matrix.}

\item{variable_names}{vector of variable names.}

\item{fomula}{the model fomula used for sem, should be a list.}
}
\description{
Get Quantile Plots from a Gaussia SEM
}
\examples{
data('example_data')
assem <- example_data
assem<- as.data.frame(assem)
t1 <- as.data.frame(apply(assem, 2, scale))

data('example_edges')
df2_tr <- example_edges # edge list
edges <- as.matrix(df2_tr)
dag_tr <- graph_from_edgelist(edges, directed = TRUE)

adjacency_matrix <- get.adjacency(dag_tr, sparse = FALSE)
variable_names <- V(dag_tr)$name
fomula <- createModel(adjacency_matrix, variable_names,'list')

model_string <- createModel(adjacency_matrix, variable_names,'string')
fitted_model1 <- sem(model_string, data = t1,std.lv = TRUE)

qplot_sem(fitted_model1,t1,adjacency_matrix,variable_names,fomula)


}
